requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
jinja2==3.1.4
python-multipart==0.0.9
